+++ CHANGES.rst
@@ -1,6 +1,8 @@
 Unreleased
 ----------
 
+- Bugfix: Handle custom scalar that subclass :class:`py_gql.schema.ScalarType` during execution.
+
 0.1.0
 -----